[FIX] account: manual reconciliation with writeoff and currency was wrong
When trying to reconcile the following line with the manual reconciliation widget
(company currency in EUR)

Debit Credit Amount_currency Date        Currency
13.11    0          0        10/26/2018  /
0       34.16      -39       11/09/2018  USD

And a write off with the following value:
Debit: 21.05

The write-off created was not with the correct debit and the entries were not totally reconciled.

The reason being that the method was converting the 21.05 with the USD rate of today and changed the debit.

We should never convert debit/credit as debit/credit in the manual reconciliation widget is always displayed in company currency. However if all entries are with the same currency, the writeoff should also have that currency set.
This commit fixes this behavior.