[FIX] web: allow string attribute on non stored fields in pivot view
Support for string attribute on fields in pivot view was added in 7ef277ec892d566
In some cases, the field is not stored and read_group method is
overridden to allow its use in a pivot view. Such a field does not
appear in the measures dict trying to set the string attribute, leading
to a client crash.

With this commit, the string attribute is set after the non stored
fields were added to the dict.

closes odoo/odoo#29393