[FIX] product: fix pricelist view on mobile phone
opw-1914133

The list of pricing in the product view was generate by default on
mobile phone. It only displayed a list with the product name.
After the commit, a kanban view is displayed. The view shows the
same informations on the mobile phone than on PC.
The princinglistitem's form view was adapted for a mobile phone use.

closes odoo/odoo#29429